Ignore dirty pyyaml submodule
When used, pyyaml generates *.pyc files, which will cause the submodule
to appear as dirty. This will bubble all the way up to any parent
repository that include Dotbot as a submodule. This change makes it so
that changes in the pyyaml repository do not cause the parent repository
to appear to have modified files.